+++ src/lmcontrol/nn/clf.py
@@ -30,6 +30,7 @@ from typing import Type, Union, List, Optional, Callable, Any
 from optuna.integration import PyTorchLightningPruningCallback
 
 from ..utils import get_logger
+from ..data_utils import load_npzs, encode_labels
 from .dataset import LMDataset, get_transforms as _get_transforms
 from lmcontrol.nn.resnet import _resnet  
 
@@ -53,143 +54,6 @@ def resnet(*, weights=None, progress=True, block=None, layers=None, planes=None,
         :members:
     """
     return _resnet(block=block, layers=layers, planes=planes, num_classes=num_classes, weights=weights, progress=progress, no_classifier=no_classifier)
-
-def encode_labels(labels, return_classes=True):
-    """This is a wrapper for sklearn.preprocessing.LabelEncoder"""
-    enc = LabelEncoder().fit(labels)
-    if return_classes:
-        return enc.transform(labels), enc.classes_
-    else:
-        return enc.transform(labels)
-    
-import numpy as np
-
-def load_npzs(npzs, logger, n=None, label_types=None):
-    """Load data from NPZ files generated from lmcontrol crop command"""
-    masks = []
-    images = []
-    paths = []
-    metadata = dict()
-
-    for npz_path in npzs:
-        logger.debug(f"Reading {npz_path}")
-        npz = np.load(npz_path)
-        
-        total_samples = len(npz['masks'])
-        
-        if n is not None and total_samples > n:
-            indices = np.random.permutation(total_samples)[:n]
-        else:
-            indices = np.arange(total_samples)
-        
-        masks.append(npz['masks'][indices])
-        images.append(npz['images'][indices])
-        paths.append(npz['paths'][indices])
-
-        md_keys = set(npz.keys()) - {'paths', 'masks', 'images'}
-        logger.debug(f"Found the following keys in {npz_path}: {' '.join(sorted(md_keys))}")
-        
-        for k in sorted(md_keys):
-            if npz[k].ndim == 0:
-                metadata.setdefault(k, []).extend([str(npz[k])] * len(indices))
-            else:
-                metadata.setdefault(k, []).extend(np.array(npz[k])[indices])
-
-    logger.debug("Merging masks")
-    masks = np.concatenate(masks, axis=0)
-    logger.debug("Merging images")
-    images = np.concatenate(images, axis=0)
-    logger.debug("Merging paths")
-    paths = np.concatenate(paths, axis=0)
-    
-    metadata = {k: np.array(v) for k, v in metadata.items()}
-
-    target_len = len(masks)
-    for k in metadata.keys():
-        if len(metadata[k]) != target_len:
-            logger.critical(f"Metadata '{k}' not found in all NPZ files")
-            raise ValueError(f"Metadata '{k}' length mismatch: expected {target_len}, got {len(metadata[k])}")
-    return masks, images, paths, metadata
-
-
-class LMDataset(Dataset):
-
-    def __init__(self, npzs, use_masks=False, return_labels=False, logger=None, transform=None, label_types=None, n=None, no_classifier=None, split=None, val_size=None, seed=None):
-        """
-        Args:
-            npzs (array-like)       : A list or tuple of paths to NPZ files containing cropped images
-        """
-        if not isinstance(npzs, (list, tuple, np.ndarray, torch.Tensor)):
-            raise ValueError(f"Got unexpected type ({type(npzs)}) for argument 'npzs'. Must be an array-like")
-        elif len(npzs) == 0:
-            raise ValueError("Got empty array-like for argument 'npzs'")
-        logger = logger or get_logger('warning')
-
-        masks, images, paths, metadata = load_npzs(npzs, logger,n,label_types)
-        if use_masks:
-            self.data = masks
-        else:
-            self.data = images
-        self.data = torch.from_numpy(self.data)[:, None, :, :]
-        self.paths = tuple(paths)
-        self.transform = transform
-
-        if not no_classifier:
-            if not isinstance(label_types, (tuple, list)):
-                label_types = [label_types]
-
-        self.labels = None
-        self.label_classes = None
-        self.label_types = None
-
-        if return_labels:
-            tmp = list()
-            self.label_classes = list()
-            self.label_types = list()
-            for k in metadata:
-                if not no_classifier:           
-                    if k not in label_types:
-                        continue
-                self.label_types.append(k)
-                labels, classes = encode_labels(metadata[k])
-                self.label_classes.append(classes)
-                tmp.append(labels)
-            self.labels = torch.from_numpy(np.stack(tmp, axis=1))
-        
-        if val_size:
-            self._split_data(split, val_size, seed)
-
-
-    def _split_data(self, split, val_size, seed):
-
-        num_samples = len(self.data)
-        indices = np.arange(num_samples)
-    
-        train_indices, val_indices = train_test_split(indices, test_size=val_size, random_state=seed, stratify=self.labels)
-
-        if split == 'train':
-            self.data = self.data[train_indices]
-            if self.labels is not None:
-                self.labels = self.labels[train_indices]
-        elif split == 'validate':
-            self.data = self.data[val_indices]
-            if self.labels is not None:
-                self.labels = self.labels[val_indices]
-
-    def __getitem__(self, i):
-        ret = self.data[i]
-        if self.transform is not None:
-            ret = self.transform(ret)
-        labels = -1 if self.labels is None else self.labels[i]
-        return ret, labels
-
-    def __len__(self):
-        return len(self.data)
-
-    @staticmethod
-    def index_to_filename(dataset, i):
-        return dataset.paths[i]
-
 
 
 class LightningResNet(L.LightningModule):
